LayoutTests/fast/mediastream/MediaStreamTrackList.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Tests MediaStreamTrackList callbacks.");

var stream1;
var stream2;

function error() {
    testFailed('Stream generation failed.');
    finishJSTest();
}

function getUserMedia(dictionary, callback) {
    try {
        navigator.webkitGetUserMedia(dictionary, callback, error);
    } catch (e) {
        testFailed('webkitGetUserMedia threw exception :' + e);
        finishJSTest();
    }
}

function onRemoveTrack(e) {
    testPassed('Remove track callback succeeded.');

    // Now test that add failes when the parent stream has been stopped.
    stream1.stop();
    shouldThrow('stream1.addTrack(stream2.getVideoTracks()[0])');
    shouldThrow('stream1.removeTrack(stream2.getVideoTracks()[0])');

    finishJSTest();
}

function onAddTrack(e) {
    testPassed('Add track callback succeeded.');
    event = e;

    stream1.onremovetrack = onRemoveTrack;
    try {
        stream1.removeTrack(event.track);
    } catch (exception) {
        testFailed("remove threw an exception.");
        finishJSTest();
    }
}

function gotStream2(s) {
    stream2 = s;
    stream1.onaddtrack = onAddTrack;

    track = stream1.getVideoTracks()[0];
    shouldBe('stream1.getTrackById(track.id).id', 'track.id');
    shouldBeNull('stream2.getTrackById(track.id)');

    try {
        stream1.addTrack(stream2.getVideoTracks()[0]);
    } catch (exception) {
        testFailed("add threw an exception.");
        finishJSTest();
    }
}

function gotStream1(s) {
    stream1 = s;

    getUserMedia({audio:true, video:true}, gotStream2);
}

getUserMedia({audio:true, video:true}, gotStream1);

window.jsTestIsAsync = true;
window.successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
