LayoutTests/fast/mediastream/MediaStreamTrackList-expected.txt

Tests MediaStreamTrackList callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stream1.getTrackById(track.id).id is track.id
PASS stream2.getTrackById(track.id) is null
PASS Add track callback succeeded.
PASS Remove track callback succeeded.
PASS stream1.addTrack(stream2.getVideoTracks()[0]) threw exception Error: InvalidStateError: DOM Exception 11.
PASS stream1.removeTrack(stream2.getVideoTracks()[0]) threw exception Error: InvalidStateError: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
